+++ Brunel_v47r2p1/GpuManager/GpuIpc/src/SocketServer.cpp
@@ -29,7 +29,7 @@ SocketServer::~SocketServer() {
 void SocketServer::readBytes(void * data, size_t size) {
   size_t total = 0u;
   while (total < size) {
-    int received = read(m_socket, (uint8_t*)data + total, size - total);
+    ssize_t received = read(m_socket, (uint8_t*)data + total, size - total);
     if (received == 0)
       throw EofException();
     if (received < 0)
@@ -41,7 +41,7 @@ void SocketServer::readBytes(void * data, size_t size) {
 void SocketServer::writeBytes(const void * data, size_t size) {
   size_t total = 0u;
   while (total < size) {
-    int sent = write(m_socket, (uint8_t*)data + total, size - total);
+    ssize_t sent = write(m_socket, (uint8_t*)data + total, size - total);
     if (sent < 0)
       throw SystemException("Write error.");
     total += static_cast<size_t>(sent);
